Move RefreshGuideScheduledTask to Guide folder
